refactor(shared): move types into shared directory
