+++ shared/types/Favorite.ts
