+++ shared/types/SharedState.ts
